:recycle: refactor: fix imports name and args in controllers
